Update type column t_id

+++ statistics_service/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, DateTime, Float, ForeignKey
+from sqlalchemy import Column, BigInteger, String, Date, DateTime, Float, ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import date, datetime
 from .database import Base
@@ -6,7 +6,7 @@ from .database import Base
 class UserProgress(Base):
     __tablename__ = "user_progress"
     
-    user_telegram_id = Column(Integer, primary_key=True, index=True)
+    user_telegram_id = Column(BigInteger, primary_key=True, index=True)
     user_name = Column(String, nullable=True)
     start_date = Column(Date, nullable=False, default=date.today)
     last_accessed_date = Column(Date, nullable=False, default=date.today)
@@ -15,9 +15,9 @@ class UserProgress(Base):
 class LessonCompletion(Base):
     __tablename__ = "lesson_completions"
     
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_telegram_id = Column(Integer, ForeignKey("user_progress.user_telegram_id"), nullable=False)
-    lesson_day = Column(Integer, nullable=False)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_telegram_id = Column(BigInteger, ForeignKey("user_progress.user_telegram_id"), nullable=False)
+    lesson_day = Column(BigInteger, nullable=False)
     completed_at = Column(DateTime, default=datetime.utcnow)
     evaluation_score = Column(Float, nullable=True)
     user = relationship("UserProgress", back_populates="completed_lessons")
